move backend into the backend.go
